Move search_multiple to Generic Search module
